Add static helper library to use on WIN32 when building Parquet unit tests to get access to inaccessible symbols [skip ci]
